+++ src/lazograph/__init__.py
@@ -1,4 +1,3 @@
 """LazoGraph product interfaces."""
 
-__version__ = "0.4.0"
-
+__version__ = "0.5.0"